requirements.txt
streamlit
requests
Pillow
pyttsx3
groq
PyPDF2